--- keras/data/__init__.py
@@ -1,2 +1,0 @@
-from .dataset import Dataset
-from .enqueuers import OrderedEnqueuer, GeneratorEnqueuer
