+++ setup.py
@@ -1,6 +1,7 @@
 from setuptools import setup
 from setuptools import find_packages
 
+
 setup(name='Keras',
       version='2.0.4',
       description='Deep Learning for Python',
